requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
DateTime==5.0
Deprecated==1.2.13
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
numpy==1.24.2
pandas==1.5.3
pykrx==1.0.40
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
wrapt==1.15.0
xlrd==2.0.1
zope.interface==5.5.2
